admin/pyproject.toml
[project]
name = "lavalink-admin"
version = "0.1.0"
description = "Admin panel for Lavalink audio node management"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "ruamel.yaml>=0.18.0",
    "docker>=7.0.0",
    "apscheduler>=3.10.0",
    "pydantic-settings>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx",  # for TestClient
]
